+++ order/entity/getmultipleorderitemsdataresponseentity.go
@@ -6,8 +6,8 @@ import (
 
 type GetMultipleOrderItemsDataResponseEntity struct{
     OrderItems	[]GetMultipleOrderItemsOrderItemsResponseEntity	`json:"order_items"`
-    OrderNumber	int	`json:"order_number"`
-    OrderId	int	`json:"order_id"`
+    OrderNumber	int64	`json:"order_number"`
+    OrderId	int64	`json:"order_id"`
 }
 func (g GetMultipleOrderItemsDataResponseEntity) String() string {
     return lib.ObjectToString(g)
