+++ order/entity/setstatustopackedbymarketplaceorderitemsresponseentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type SetStatusToPackedByMarketplaceOrderItemsResponseEntity struct{
-    OrderItemId	int	`json:"order_item_id"`
-    PurchaseOrderId	int	`json:"purchase_order_id"`
+    OrderItemId	int64	`json:"order_item_id"`
+    PurchaseOrderId	int64	`json:"purchase_order_id"`
     PurchaseOrderNumber	string	`json:"purchase_order_number"`
     PackageId	string	`json:"package_id"`
     ShipmentProvider	string	`json:"shipment_provider"`
