+++ returnandrefund/entity/getreverseorderdetaildataresponseentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type GetReverseOrderDetailDataResponseEntity struct{
-    ReverseOrderId	int	`json:"reverse_order_id"`
-    TradeOrderId	int	`json:"trade_order_id"`
+    ReverseOrderId	int64	`json:"reverse_order_id"`
+    TradeOrderId	int64	`json:"trade_order_id"`
     RequestType	string	`json:"request_type"`
     ShippingType	string	`json:"shipping_type"`
     IsRtm	bool	`json:"is_rtm"`
