+++ returnandrefund/entity/getreverseordersforsellerbuyerresponseentity.go
@@ -5,7 +5,7 @@ import (
 )
 
 type GetReverseOrdersForSellerBuyerResponseEntity struct{
-    BuyerId	int	`json:"buyer_id"`
+    BuyerId	int64	`json:"buyer_id"`
 }
 func (g GetReverseOrdersForSellerBuyerResponseEntity) String() string {
     return lib.ObjectToString(g)
