+++ returnandrefund/entity/getreverseordersforselleritemsresponseentity.go
@@ -5,8 +5,8 @@ import (
 )
 
 type GetReverseOrdersForSellerItemsResponseEntity struct{
-    ReverseOrderId	int	`json:"reverse_order_id"`
-    TradeOrderId	int	`json:"trade_order_id"`
+    ReverseOrderId	int64	`json:"reverse_order_id"`
+    TradeOrderId	int64	`json:"trade_order_id"`
     RequestType	string	`json:"request_type"`
     IsRtm	bool	`json:"is_rtm"`
     ShippingType	string	`json:"shipping_type"`
